integration/custom/deepseek-v3-ascend310p/src/requirements.txt
pyyaml
msgspec
requests
protobuf
numpy==1.26
transformers
decorator
scipy
